Zmieniono nazwę projektu na Schowek.Client.WebAPI
