+++ Schowek.Client.WebAPI/Schowek.Client.WebAPI.csproj
@@ -6,6 +6,7 @@
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
+  <PropertyGroup Condition=" '$(RunConfiguration)' == 'SchowekAPI' " />
   <ItemGroup>
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.1" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.8" />
@@ -22,7 +23,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Schowek.Library\Schowek.Library.csproj" />
+    <ProjectReference Include="..\Schowek.Shared.Core\Schowek.Shared.Core.csproj" />
   </ItemGroup>
 
 </Project>
