+++ Schowek.Client.WebAPI/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DevelopmentConnection": "DataSource=../Schowek.Library/Data/app.db;Cache=Shared"
+    "DevelopmentConnection": "DataSource=../Schowek.Shared.Core/Data/app.db;Cache=Shared"
   },
   "Logging": {
     "LogLevel": {
